+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # CHANGELOG
 
+## 0.1.2 - 2020-11-20
+- Add support for target properties.
+- Add installs.
+- Require cmake >= 3.15.
+- Support settings and caching settings.
+- Support config when running cmkr build.
+
 ## 0.1.1 - 2020-11-19
 - Add support for globbing.
 - Add support for find_package components.
 - Add options.
-- Add installs.
 - Support aliases.
 - Support interface libs (header-only libs).
 - Support testing.
-- Require cmake >= 3.15.